requirements.txt
Django>=4.0.1,<4.1
djangorestframework-gis>=1.0.0,<1.1
django-leaflet>=0.29.0,<0.30
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.22.1,<0.23
Pillow>=10.0.1,<11.0.0
uwsgi>=2.0.20,<2.1
